requirements/base.txt
appdirs==1.4.3
argon2-cffi==16.3.0
cffi==1.10.0
click==6.7
coreapi==2.3.0
coreapi-cli==1.0.6
coreschema==0.0.4
decorator==4.0.11
Django==1.10.6
django-appconf==1.0.2
django-debug-toolbar==1.6
django-extensions==1.7.4
django-imagekit==4.0
django-rest-swagger==2.1.2
djangorestframework==3.6.2
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.1
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
pilkit==2.0
Pillow==4.0.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
pycparser==2.17
Pygments==2.2.0
pyparsing==2.2.0
requests==2.13.0
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
uritemplate==3.0.0
wcwidth==0.1.7
gunicorn==19.7.1
